PhotoTaker show selected item image
